project-Chatbot/requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-community>=0.1.2
langchain-core>=0.1.4
streamlit>=1.29.0
python-dotenv>=1.0.0
tiktoken>=0.5.2
openai>=1.12.0 